pyproject.toml
[project]
name = "fastapi-rbac"
version = "0.1.0"
description = "FastAPI Role Basi"
authors = [
    {name = "Gabriel Martin del Campo y Eternod", email = "gmartindelcye@gmail.com"},
]
dependencies = [
    "fastapi>=0.101.1",
    "httpx>=0.24.1",
    "uvicorn>=0.23.2",
    "pymongo>=4.4.1",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
